Move AmortizedIStreamReader to driver/utils/amortized_istream_reader.h Move ObjectPool to driver/utils/object_pool.h Move resize_without_initialization() to driver/utils/resize_without_initialization.h
